3/08/22
On this day, I have decided to change the entirety of my html. This is because I realised that if I still continue I will encounter the same problem again, which I don’t want to do.

Therefore, I have changed my HTML code to something more efficient and much more clearer. I have also changed the layout of the HTML code.

On this new version, I have gotten rid of the “main-container” to line up all the element that are inside the div. Instead I have used a div called “wrapper” this div is going to be added to each section, so that the elements that are inside this div will be aligned perfectly with other elements that have the class of “wrapper” as well.
I have also changed the layout of my code, in the first version, the hero section texts comes before the navigation bar. However now in this new version, the navigation bar comes before the hero section texts. I think this was one of the problems that I have encountered which made the next section, still show in the header instead of showing beneath it.