3/08/22
On this day, I have made the style and also how the fixed navigation bar will look like when it is scrolled. Please refer to my log book to see the details.

To make things short, I have made a JavaScript file which pretty much adds a new class to the navigation bar, this new class then includes the attributes or code on how the navigation bar will look like when the user scrolls.