Minor Changes on the map javascript - Based on the feedbacks I have received
On this day, I have made some changes on the code of the website especially in the way -the map works. This is because I have received a feedback from Mr Swanink, regarding to how there are many overlapping when a card is open. It should be that, when one card is open and you press on a marker, the previous the card that was opened should be closed.

And that is what I have done on this day, coded a feature, where if a user clicks on a new marker, if there is open card ot woo;; close it and open the card that corresponds to marker that was clicked